Version final del proyecto *El sistema muestra una lista de cursos limitada *dentro de cada curso muestra los alumnos junto con sus asignaturas y las calificaciones que sacaron en cada una de las 5 pruebas de cada uno con valores entre 0.0 y 5.0

+++ Entidades/Evaluaciones.cs
@@ -12,10 +12,9 @@ namespace CoreEscuela.Entidades
         public string Nombre { get; set; }
 
         public Alumno Alumno { get; set; }
-
         public Asignatura Asignatura { get; set; }
 
-        public float Nota { get; set; }
+        public double Nota { get; set; }
 
         public Evaluaciones() => UniqueId = Guid.NewGuid().ToString();
     }
